requirements.txt
fastapi
uvicorn[standard]
requests
anthropic
python-dotenv
twilio
